+++ UnitTests/PopupBridge_UnitTests.m
@@ -13,6 +13,7 @@
 @implementation MockUserContentController
 
 NSString * const kPOPScriptMessageHandlerName = @"POPPopupBridge";
+NSString * const kReturnURLScheme = @"com.braintreepayments.popupbridgeexample";
 
 - (void)addScriptMessageHandler:(id<WKScriptMessageHandler>)scriptMessageHandler name:(NSString *)name {
     self.scriptMessageHandler = scriptMessageHandler;
@@ -28,12 +29,6 @@ NSString * const kPOPScriptMessageHandlerName = @"POPPopupBridge";
 
 static void (^webviewReadyBlock)(void);
 
-- (void)setUp {
-    [super setUp];
-
-    [POPPopupBridge setReturnURLScheme:@"com.braintreepayments.popupbridgeexample"];
-}
-
 - (void)tearDown {
     webviewReadyBlock = nil;
 }
@@ -44,7 +39,7 @@ static void (^webviewReadyBlock)(void);
 
     XCTAssertEqual(webView.configuration.userContentController.userScripts.count, 0);
 
-    __unused POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:webView delegate:delegate];
+    __unused POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:webView urlScheme:kReturnURLScheme delegate:delegate];
 
     XCTAssertEqual(webView.configuration.userContentController.userScripts.count, 1);
     WKUserScript *userScript = webView.configuration.userContentController.userScripts[0];
@@ -57,30 +52,11 @@ static void (^webviewReadyBlock)(void);
     MockUserContentController *mockUserContentController = [[MockUserContentController alloc] init];
     configuration.userContentController = mockUserContentController;
     WKWebView *webView = [[WKWebView alloc] initWithFrame:CGRectZero configuration:configuration];
-
-    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:webView delegate:OCMProtocolMock(@protocol(POPPopupBridgeDelegate))];
+    
+    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:webView urlScheme:kReturnURLScheme delegate:OCMProtocolMock(@protocol(POPPopupBridgeDelegate))];
         
     XCTAssertEqual(mockUserContentController.scriptMessageHandler, pub);
     XCTAssertEqualObjects(mockUserContentController.name, kPOPScriptMessageHandlerName);
-}
-
-- (void)testInit_whenSchemeIsNotSet_throwsError {
-#pragma clang diagnostic push
-#pragma clang diagnostic ignored "-Wnonnull"
-    [POPPopupBridge setReturnURLScheme:nil];
-#pragma clang diagnostic pop
-    WKWebView *webView = [[WKWebView alloc] initWithFrame:CGRectZero];
-
-    NSException *thrownException;
-    POPPopupBridge *popupBridge;
-    @try {
-        popupBridge = [[POPPopupBridge alloc] initWithWebView:webView delegate:(id<POPPopupBridgeDelegate>)[[NSObject alloc] init]];
-    } @catch (NSException *exception) {
-        thrownException = exception;
-    } @finally {
-        XCTAssertEqualObjects(thrownException.name, @"POPPopupBridgeSchemeNotSet");
-        XCTAssertNil(popupBridge);
-    }
 }
 
 - (void)testReceiveScriptMessage_whenMessageContainsURL_requestsPresentationOfSafariViewController {
@@ -91,7 +67,7 @@ static void (^webviewReadyBlock)(void);
     OCMStub([stubWebView configuration]).andDo(nil);
     id mockDelegate = OCMProtocolMock(@protocol(POPPopupBridgeDelegate));
 
-    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:stubWebView delegate:mockDelegate];
+    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:stubWebView urlScheme:kReturnURLScheme delegate:mockDelegate];
     [pub userContentController:[[WKUserContentController alloc] init] didReceiveScriptMessage:stubMessage];
 
     OCMVerify([mockDelegate popupBridge:pub requestsPresentationOfViewController:[OCMArg checkWithBlock:^BOOL(UIViewController *viewController) {
@@ -107,7 +83,7 @@ static void (^webviewReadyBlock)(void);
     OCMStub([stubWebView configuration]).andDo(nil);
     id mockDelegate = OCMProtocolMock(@protocol(POPPopupBridgeDelegate));
 
-    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:stubWebView delegate:mockDelegate];
+    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:stubWebView urlScheme:kReturnURLScheme delegate:mockDelegate];
     [pub userContentController:[[WKUserContentController alloc] init] didReceiveScriptMessage:stubMessage];
 
     OCMVerify([mockDelegate popupBridge:pub willOpenURL:[NSURL URLWithString:@"http://example.com/?hello=world"]]);
@@ -121,7 +97,7 @@ static void (^webviewReadyBlock)(void);
     OCMStub([stubWebView configuration]).andDo(nil);
     id mockDelegate = OCMStrictProtocolMock(@protocol(POPPopupBridgeDelegate));
 
-    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:stubWebView delegate:mockDelegate];
+    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:stubWebView urlScheme:kReturnURLScheme delegate:mockDelegate];
     [pub userContentController:[[WKUserContentController alloc] init] didReceiveScriptMessage:stubMessage];
 }
 
@@ -133,7 +109,7 @@ static void (^webviewReadyBlock)(void);
     OCMStub([stubWebView configuration]).andDo(nil);
     id mockDelegate = OCMStrictProtocolMock(@protocol(POPPopupBridgeDelegate));
 
-    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:stubWebView delegate:mockDelegate];
+    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:stubWebView urlScheme:kReturnURLScheme delegate:mockDelegate];
     [pub userContentController:[[WKUserContentController alloc] init] didReceiveScriptMessage:stubMessage];
 }
 
@@ -145,7 +121,7 @@ static void (^webviewReadyBlock)(void);
     WKWebView *webView = OCMClassMock([WKWebView class]);
     SFSafariViewController *stubSafari = [[SFSafariViewController alloc] initWithURL:[NSURL URLWithString:@"http://example.com"]];
     
-    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:webView delegate:OCMProtocolMock(@protocol(POPPopupBridgeDelegate))];
+    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:webView urlScheme:kReturnURLScheme delegate:OCMProtocolMock(@protocol(POPPopupBridgeDelegate))];
     [pub userContentController:[[WKUserContentController alloc] init] didReceiveScriptMessage:message];
 
     [((id <SFSafariViewControllerDelegate>)pub) safariViewControllerDidFinish:stubSafari];
@@ -164,7 +140,7 @@ static void (^webviewReadyBlock)(void);
     OCMStub(message.name).andReturn(kPOPScriptMessageHandlerName);
     WKWebView *webView = OCMClassMock([WKWebView class]);
 
-    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:webView delegate:OCMProtocolMock(@protocol(POPPopupBridgeDelegate))];
+    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:webView urlScheme:kReturnURLScheme delegate:OCMProtocolMock(@protocol(POPPopupBridgeDelegate))];
     [pub userContentController:[[WKUserContentController alloc] init] didReceiveScriptMessage:message];
     [POPPopupBridge openURL:[NSURL URLWithString:@"com.braintreepayments.popupbridgeexample://popupbridgev1/return?something=foo&other=bar"]];
 
@@ -185,7 +161,7 @@ static void (^webviewReadyBlock)(void);
     OCMStub(message.name).andReturn(kPOPScriptMessageHandlerName);
     WKWebView *webView = OCMClassMock([WKWebView class]);
 
-    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:webView delegate:OCMProtocolMock(@protocol(POPPopupBridgeDelegate))];
+    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:webView urlScheme:kReturnURLScheme delegate:OCMProtocolMock(@protocol(POPPopupBridgeDelegate))];
     [pub userContentController:[[WKUserContentController alloc] init] didReceiveScriptMessage:message];
     [POPPopupBridge openURL:[NSURL URLWithString:@"com.braintreepayments.popupbridgeexample://popupbridgev1/return"]];
 
@@ -206,7 +182,7 @@ static void (^webviewReadyBlock)(void);
     OCMStub(message.name).andReturn(kPOPScriptMessageHandlerName);
     WKWebView *webView = OCMClassMock([WKWebView class]);
 
-    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:webView delegate:OCMProtocolMock(@protocol(POPPopupBridgeDelegate))];
+    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:webView urlScheme:kReturnURLScheme delegate:OCMProtocolMock(@protocol(POPPopupBridgeDelegate))];
     [pub userContentController:[[WKUserContentController alloc] init] didReceiveScriptMessage:message];
     [POPPopupBridge openURL:[NSURL URLWithString:@"com.braintreepayments.popupbridgeexample://popupbridgev1/return#something=foo&other=bar"]];
 
@@ -226,7 +202,7 @@ static void (^webviewReadyBlock)(void);
     OCMStub(message.name).andReturn(kPOPScriptMessageHandlerName);
     WKWebView *webView = OCMClassMock([WKWebView class]);
 
-    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:webView delegate:OCMProtocolMock(@protocol(POPPopupBridgeDelegate))];
+    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:webView urlScheme:kReturnURLScheme delegate:OCMProtocolMock(@protocol(POPPopupBridgeDelegate))];
     [pub userContentController:[[WKUserContentController alloc] init] didReceiveScriptMessage:message];
     [POPPopupBridge openURL:[NSURL URLWithString:@"com.braintreepayments.popupbridgeexample://popupbridgev1/return"]];
 
@@ -248,7 +224,7 @@ static void (^webviewReadyBlock)(void);
     WKWebView *webView = OCMStrictClassMock([WKWebView class]);
     OCMStub([webView configuration]).andReturn(nil);
 
-    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:webView delegate:OCMProtocolMock(@protocol(POPPopupBridgeDelegate))];
+    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:webView urlScheme:kReturnURLScheme delegate:OCMProtocolMock(@protocol(POPPopupBridgeDelegate))];
     [pub userContentController:[[WKUserContentController alloc] init] didReceiveScriptMessage:message];
 
     BOOL result = [POPPopupBridge openURL:[NSURL URLWithString:@"not.the.right.scheme://popupbridgev1/return?something=foo"]];
@@ -263,7 +239,7 @@ static void (^webviewReadyBlock)(void);
     WKWebView *webView = OCMStrictClassMock([WKWebView class]);
     OCMStub([webView configuration]).andReturn(nil);
 
-    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:webView delegate:OCMProtocolMock(@protocol(POPPopupBridgeDelegate))];
+    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:webView urlScheme:kReturnURLScheme delegate:OCMProtocolMock(@protocol(POPPopupBridgeDelegate))];
     [pub userContentController:[[WKUserContentController alloc] init] didReceiveScriptMessage:message];
 
     BOOL result = [POPPopupBridge openURL:[NSURL URLWithString:@"com.braintreepayments.popupbridgeexample://notcorrect/return?something=foo"]];
@@ -274,7 +250,7 @@ static void (^webviewReadyBlock)(void);
     id mockDelegate = OCMProtocolMock(@protocol(POPPopupBridgeDelegate));
     WKWebView *webView = [[WKWebView alloc] initWithFrame:CGRectZero];
     webView.navigationDelegate = self;
-    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:webView delegate:mockDelegate];
+    POPPopupBridge *pub = [[POPPopupBridge alloc] initWithWebView:webView urlScheme:kReturnURLScheme delegate:mockDelegate];
 
     id expectation = [self expectationWithDescription:@"Called JS"];